requirements.txt
asgiref==3.5.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Delorean==1.0.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
docutils==0.18.1
drf-yasg==1.20.0
humanize==4.0.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-gettext==4.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
django-phonenumber-field[phonenumbers]